scripts/requirements.txt
"""Requirements for seed scripts."""

sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
